provide some help with obscure shell errors
I hit both of these and had to noodle in the source to figure out what was up.
I'm not sure why fastlane didn't ask me about setting up account details (it has in other contexts). Perhaps because I'm only using it to run notarize and it doesn't have a bunch of other app tasks.